pyproject.toml
[project]
name = "youth-event-registration-app"
version = "0.1.0"
description = ""
authors = [
    {name = "joaobarreto27",email = "joao.vito1951@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "python-jose[cryptography] (==3.3.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "streamlit (>=1.53.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
